moved common platform code into shared are for win32 projects

+++ source/shared_lib/sources/platform/sdl/window.cpp
@@ -35,7 +35,7 @@ namespace Shared{ namespace Platform{
 // Matze: hack for now...
 static Window* global_window = 0;
 static int oldX=0,oldY=0;
-unsigned int Window::lastMouseEvent = 0;	/** for use in mouse hover calculations */
+int64 Window::lastMouseEvent = 0;	/** for use in mouse hover calculations */
 Vec2i Window::mousePos;
 MouseState Window::mouseState;
 bool Window::isKeyPressedDown = false;
